Change build to Python 2.7
- Workaround for #14

Currently there is a yet to be found numerical issue when building under Python3.x which is
not related to a dependency version change or division as far as I can tell

There is also a parsing error where ALBUS rejects stations other than the first station on the
stations list. It is possible a change in behaviour with iterators or a parsing issue which
I am yet to trace down.

These are two distinct unrelated problems